refactor(cmd): split checkin.go into per-concern files
cmd/checkin.go was 1680 LoC mixing command wiring, rendering paths, helper
functions, and the review-round display. This pure structural refactor splits
it into 6 focused files, all in package main:

  checkin.go         — command wiring + top-level dispatch (~180 LoC)
  checkin_turns.go   — renderCheckinTurns (interleaved-turn primary view)
  checkin_raw.go     — raw event rendering + proxy path
  checkin_review.go  — review-round summary paths
  checkin_list.go    — no-arg session list
  checkin_tools.go   — payload-parsing and result-summary helpers
  checkin_legacy.go  — screen-scrape fallback

No behaviour changes. go build and go test pass.
